Create new file for NewItemForm
